livekit-voice-integration/requirements.txt
livekit==0.11.0
livekit-agents==0.8.0
openai-whisper==20231117
openai==1.35.0
numpy==1.24.3
torch==2.0.1
torchaudio==2.0.2
fastapi==0.104.1
uvicorn==0.24.0
websockets==11.0.3
python-multipart==0.0.6
